Move config to it's own file
