+++ gTile@shuairan/src/3_8/ui/GridSettingsButton.ts
@@ -1,4 +1,5 @@
-import { preferences, refreshGrids } from "../extension";
+import { preferences } from "../config";
+import { refreshGrids } from "../extension";
 
 const St = imports.gi.St;
